requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-emit-index-url
#
asgiref==3.7.2
    # via django
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via requests
diff-match-patch==20230430
    # via django-reversion-compare
dj-database-url==0.5.0
    # via -r requirements.in
django-reversion-compare==0.14.0
    # via -r requirements.in
django-reversion==3.0.9
    # via
    #   -r requirements.in
    #   django-reversion-compare
django==3.1.14
    # via
    #   -r requirements.in
    #   django-reversion
    #   djangorestframework
djangorestframework==3.14.0
    # via -r requirements.in
idna==3.4
    # via requests
phonenumberslite==8.12.18
    # via -r requirements.in
psycopg2-binary==2.8.6
    # via -r requirements.in
pytz==2023.3.post1
    # via
    #   django
    #   djangorestframework
requests==2.31.0
    # via -r requirements.in
sqlparse==0.4.4
    # via django
typing-extensions==4.7.1
    # via asgiref
urllib3==2.0.4
    # via requests
